feat: displaying expenses, adding expenses
