+++ src/app/loading.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import Loader from "@/components/Loader";
+
+const Loading = () => {
+  return <Loader />;
+};
+
+export default Loading;
