+++ src/components/Landing.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+type Props = {};
+
+const Landing = (props: Props) => {
+  return <div>Landing</div>;
+};
+
+export default Landing;
