+++ src/components/Loader.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+type Props = {};
+
+const Loader = (props: Props) => {
+  return <div>Loading...</div>;
+};
+
+export default Loader;
